# swagger 00.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EcoLens API",
    "description": "Sustainable lifestyle application API",
    "version": "v1"
  },
  "paths": {
    "/api/Activity/upload": {
      "post": {
        "tags": [
          "Activity"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Label",
                  "Quantity"
                ],
                "type": "object",
                "properties": {
                  "Base64Image": {
                    "type": "string"
                  },
                  "ImageFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Label": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "Quantity": {
                    "type": "number",
                    "format": "double"
                  },
                  "Category": {
                    "$ref": "#/components/schemas/CarbonCategory"
                  }
                }
              },
              "encoding": {
                "Base64Image": {
                  "style": "form"
                },
                "ImageFile": {
                  "style": "form"
                },
                "Label": {
                  "style": "form"
                },
                "Quantity": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityLogResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityLogResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityLogResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Activity/chart-data": {
      "get": {
        "tags": [
          "Activity"
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartDataPointDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartDataPointDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartDataPointDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/carbon-reference": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "label",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarbonReferenceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarbonReferenceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarbonReferenceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCarbonReferenceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCarbonReferenceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCarbonReferenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonReferenceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonReferenceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonReferenceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/carbon-reference/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/admin/users/{id}/ban": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "totalEmission": {
            "type": "number",
            "format": "double"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "BanUserRequestDto": {
        "type": "object",
        "properties": {
          "ban": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CarbonCategory": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CarbonReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "labelName": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/CarbonCategory"
          },
          "co2Factor": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChartDataPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "nullable": true
          },
          "emission": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "UpsertCarbonReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "labelName": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/CarbonCategory"
          },
          "co2Factor": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "在下方输入 JWT。示例：Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}